+++ src/fr/maxlego08/menu/loader/actions/SoundLoader.java
@@ -3,8 +3,8 @@ package fr.maxlego08.menu.loader.actions;
 import com.cryptomorin.xseries.XSound;
 import fr.maxlego08.menu.api.loader.ActionLoader;
 import fr.maxlego08.menu.api.requirement.Action;
+import fr.maxlego08.menu.api.sound.SoundOption;
 import fr.maxlego08.menu.api.utils.TypedMapAccessor;
-import fr.maxlego08.menu.requirement.actions.BroadcastSoundAction;
 import fr.maxlego08.menu.requirement.actions.SoundAction;
 import fr.maxlego08.menu.sound.ZSoundOption;
 
@@ -19,11 +19,14 @@ public class SoundLoader implements ActionLoader {
 
     @Override
     public Action load(String path, TypedMapAccessor accessor, File file) {
+        return new SoundAction(loadSound(path, accessor, file));
+    }
 
+    protected SoundOption loadSound(String path, TypedMapAccessor accessor, File file){
         String sound = accessor.getString("sound");
         float pitch = accessor.getFloat("pitch", 1f);
         float volume = accessor.getFloat("volume", 1f);
         XSound xSound = sound == null || sound.isEmpty() ? null : XSound.matchXSound(sound).orElse(null);
-        return new BroadcastSoundAction(new ZSoundOption(xSound, sound, volume, pitch, xSound == null));
+        return new ZSoundOption(xSound, sound, pitch, volume, xSound == null);
     }
 }