+++ src/fr/maxlego08/menu/save/Config.java
@@ -73,6 +73,7 @@ public class Config implements Savable {
 
     public static boolean enableCooldownClick = true;
     public static long cooldownClickMilliseconds = 100;
+    public static long cachePlaceholderAPI = 500;
 
     /**
      * static Singleton instance.