+++ changelog.md
@@ -34,6 +34,14 @@
 
 # Unreleased
 
+# 1.0.2.6
+
+- Add BlockPlaceEvent for Anti Dupe
+- Add placeholder %zmenu_test% for testing
+- Fix sound action, pitch and volume was reversed
+- Add ``job`` requirement. Check if the player has a job with JobsReborn plugins
+- Added a cache on placeholders. By default, the cache is 500ms. That corresponds to opening a menu. So if you have several times the same placeholder used, it will only be called once.
+
 # 1.0.2.5
 
 - Add new lore as string support